remove some bloat from tilted/core.rs
