+++ src/recom/tilted/core.rs
@@ -22,11 +22,15 @@ use super::super::{
     cut_edge_dist_pair, node_bound, random_split, uniform_dist_pair, RecomParams, RecomProposal,
     RecomVariant,
 };
+use super::packets::{
+    collect_tilted_results, send_tilted_jobs, stop_tilted_workers, ScoredProposal,
+    TiltedJobPacket, TiltedResultPacket, TiltedScorePacket, TiltedStatsPacket,
+};
+use super::writers::{start_tilted_score_writer, start_tilted_stats_writer};
 use crate::buffers::{
     graph_connected_buffered, ConnectivityBuffers, SpanningTreeBuffer, SplitBuffer, SubgraphBuffer,
 };
 use crate::graph::Graph;
-use crate::objectives::IncrementalObjective;
 use crate::partition::Partition;
 use crate::spanning_tree::{RMSTSampler, RegionAwareSampler, SpanningTreeSampler, USTSampler};
 use crate::stats::{ScoresWriter, SelfLoopCounts, SelfLoopReason, StatsWriter};
@@ -119,206 +123,6 @@ pub trait ScoringBackend: Send + Clone {
     /// Returns per-district scores to forward to the score writer for the
     /// just-accepted step, or `None` to keep replaying the last vector.
     fn step_district_scores(&self, state: &Self::State) -> Option<Vec<f64>>;
-}
-
-/// Saves the current state of the two affected districts before scoring a
-/// proposal. Used by [`FullRescoreBackend`] to revert after temp-applying.
-fn save_revert_buffer(partition: &Partition, proposal: &RecomProposal, revert: &mut RecomProposal) {
-    revert.a_label = proposal.a_label;
-    revert.b_label = proposal.b_label;
-    revert.a_pop = partition.dist_pops[proposal.a_label];
-    revert.b_pop = partition.dist_pops[proposal.b_label];
-    revert.a_nodes.clear();
-    revert
-        .a_nodes
-        .extend_from_slice(&partition.dist_nodes[proposal.a_label]);
-    revert.b_nodes.clear();
-    revert
-        .b_nodes
-        .extend_from_slice(&partition.dist_nodes[proposal.b_label]);
-}
-
-/// Scoring backend that rescores from scratch via a closure on every
-/// candidate. Cheaper to set up than [`IncrementalBackend`] but pays
-/// `O(districts)` per candidate rather than `O(boundary)`.
-/// Useful for prototying new objectives and for objectives
-/// that require knowledge of the full partition.
-#[derive(Clone, Copy)]
-pub struct FullRescoreBackend<F>
-where
-    F: Fn(&Graph, &Partition) -> f64 + Send + Clone + Copy,
-{
-    pub obj_fn: F,
-}
-
-impl<F> ScoringBackend for FullRescoreBackend<F>
-where
-    F: Fn(&Graph, &Partition) -> f64 + Send + Clone + Copy,
-{
-    type State = ();
-    type Scratch = RecomProposal;
-
-    fn init_state(&self, _graph: &Graph, _partition: &Partition) -> Self::State {}
-
-    fn make_scratch(&self, buf_size: usize) -> Self::Scratch {
-        RecomProposal::new_buffer(buf_size)
-    }
-
-    fn initial_score(&self, graph: &Graph, partition: &Partition, _state: &Self::State) -> f64 {
-        (self.obj_fn)(graph, partition)
-    }
-
-    fn initial_district_scores(&self, _state: &Self::State) -> Vec<f64> {
-        Vec::new()
-    }
-
-    fn score_candidate(
-        &self,
-        graph: &Graph,
-        partition: &mut Partition,
-        _state: &Self::State,
-        scratch: &mut Self::Scratch,
-        proposal: &RecomProposal,
-    ) -> f64 {
-        save_revert_buffer(partition, proposal, scratch);
-        partition.update(proposal);
-        let score = (self.obj_fn)(graph, partition);
-        partition.update(scratch);
-        score
-    }
-
-    fn apply_accepted(
-        &self,
-        _graph: &Graph,
-        partition: &mut Partition,
-        _state: &mut Self::State,
-        proposal: &RecomProposal,
-    ) {
-        // Full-rescore doesn't read dist_adj, so no benefit to keeping it warm.
-        partition.update(proposal);
-    }
-
-    fn step_district_scores(&self, _state: &Self::State) -> Option<Vec<f64>> {
-        None
-    }
-}
-
-/// Scoring backend that maintains a per-thread cached objective state and
-/// queries it directly, avoiding full rescore on every candidate.
-#[derive(Clone)]
-pub struct IncrementalBackend<O>
-where
-    O: IncrementalObjective + Send + Clone + 'static,
-    O::State: Send + Clone + 'static,
-{
-    pub objective: O,
-}
-
-impl<O> ScoringBackend for IncrementalBackend<O>
-where
-    O: IncrementalObjective + Send + Clone + 'static,
-    O::State: Send + Clone + 'static,
-{
-    type State = O::State;
-    type Scratch = ();
-
-    fn init_state(&self, graph: &Graph, partition: &Partition) -> Self::State {
-        self.objective.init(graph, partition)
-    }
-
-    fn make_scratch(&self, _buf_size: usize) -> Self::Scratch {}
-
-    fn initial_score(&self, _graph: &Graph, _partition: &Partition, state: &Self::State) -> f64 {
-        self.objective.score_state(state)
-    }
-
-    fn initial_district_scores(&self, state: &Self::State) -> Vec<f64> {
-        self.objective.district_scores(state)
-    }
-
-    fn score_candidate(
-        &self,
-        graph: &Graph,
-        _partition: &mut Partition,
-        state: &Self::State,
-        _scratch: &mut Self::Scratch,
-        proposal: &RecomProposal,
-    ) -> f64 {
-        self.objective.score_proposal(graph, state, proposal)
-    }
-
-    fn apply_accepted(
-        &self,
-        graph: &Graph,
-        partition: &mut Partition,
-        state: &mut Self::State,
-        proposal: &RecomProposal,
-    ) {
-        self.objective.apply_proposal(graph, state, proposal);
-        partition.update_with_dist_adj(proposal, graph);
-    }
-
-    fn step_district_scores(&self, state: &Self::State) -> Option<Vec<f64>> {
-        Some(self.objective.district_scores(state))
-    }
-}
-
-/// A unit of work sent from the main thread to a worker.
-struct TiltedJobPacket {
-    /// The accepted proposal to apply, or `None` if no state change.
-    diff: Option<RecomProposal>,
-    /// The current chain score, used by the worker for accept/reject decisions.
-    current_score: f64,
-    /// A sentinel used to kill the worker thread.
-    terminate: bool,
-}
-
-/// The result of one round of work from a worker.
-struct TiltedResultPacket {
-    /// Number of tilted-criterion rejections (self-loops).
-    rejections: usize,
-    /// Accepted proposals from this round.
-    proposals: Vec<ScoredProposal>,
-}
-
-/// A proposal accepted by a worker, plus the score it would produce.
-struct ScoredProposal {
-    /// Random ID used for deterministic interleaving (same technique as `run.rs`).
-    id: u64,
-    /// The proposal accepted by a worker.
-    proposal: RecomProposal,
-    /// The objective score after applying `proposal`.
-    score: f64,
-}
-
-/// A chain-statistics write packet sent from the chain thread to the writer thread.
-struct TiltedStatsPacket {
-    /// The accepted proposal step.
-    step: u64,
-    /// The accepted proposal, or `None` only for the termination sentinel.
-    proposal: Option<RecomProposal>,
-    /// Tilted rejection counts since the previous accepted proposal.
-    counts: SelfLoopCounts,
-    /// A sentinel used to stop the writer thread.
-    terminate: bool,
-}
-
-/// A score write packet sent from the chain thread to the score-writer thread.
-struct TiltedScorePacket {
-    /// First chain step represented by this packet.
-    first_step: u64,
-    /// Last chain step represented by this packet.
-    last_step: u64,
-    /// Objective score for each step in this packet.
-    score: f64,
-    /// Best score seen so far for each step in this packet.
-    best_score: f64,
-    /// Per-district scores to carry forward from this point on, if the chain
-    /// state changed. `None` on pure-rejection runs, where the writer keeps
-    /// reusing its previously cached vector.
-    district_scores: Option<Vec<f64>>,
-    /// A sentinel used to stop the writer thread.
-    terminate: bool,
 }
 
 /// Reusable worker-side buffers for one tilted worker, parameterized on the
@@ -574,139 +378,6 @@ fn fill_pair_subgraph<S>(
     dist_b: usize,
 ) {
     partition.subgraph(graph, &mut buffers.subgraph, dist_a, dist_b);
-}
-
-/// Starts a chain-statistics writer thread.
-///
-/// # Arguments
-///
-/// * `graph` - Writer-owned graph clone.
-/// * `partition` - Writer-owned starting partition, updated from accepted proposals.
-/// * `writer` - Stats writer receiving accepted proposals and self-loop counts.
-/// * `recv` - Channel receiving asynchronous stats write packets.
-fn start_tilted_stats_writer(
-    graph: &Graph,
-    mut partition: Partition,
-    writer: &mut dyn StatsWriter,
-    recv: Receiver<TiltedStatsPacket>,
-) {
-    writer.init(graph, &partition).unwrap();
-    let mut next = recv.recv().unwrap();
-    while !next.terminate {
-        if let Some(proposal) = next.proposal {
-            partition.update(&proposal);
-            writer
-                .step(next.step, graph, &partition, &proposal, &next.counts)
-                .unwrap();
-        } else {
-            writer
-                .self_loop(next.step, graph, &partition, &next.counts)
-                .unwrap();
-        }
-        next = recv.recv().unwrap();
-    }
-    writer.close().unwrap();
-}
-/// Starts a score writer thread.
-///
-/// # Arguments
-///
-/// * `writer` - Score writer receiving per-step objective scores.
-/// * `initial_score` - Objective score of the starting partition.
-/// * `initial_district_scores` - Per-district score vector for the starting
-///   partition. An empty vector switches the writer to the legacy three-column
-///   `step,score,best_score` output; a non-empty vector emits
-///   `step,score,best_score,d_0,...,d_{N-1}` and the writer caches it so
-///   pure-rejection packets can replay the last-known district scores.
-/// * `recv` - Channel receiving asynchronous score write packets.
-fn start_tilted_score_writer(
-    writer: &mut ScoresWriter,
-    initial_score: f64,
-    initial_district_scores: Vec<f64>,
-    recv: Receiver<TiltedScorePacket>,
-) {
-    writer
-        .init(initial_score, &initial_district_scores)
-        .unwrap();
-    let mut last_districts = initial_district_scores;
-    let mut next = recv.recv().unwrap();
-    while !next.terminate {
-        if let Some(new_districts) = next.district_scores.take() {
-            last_districts = new_districts;
-        }
-        for step in next.first_step..=next.last_step {
-            writer
-                .step(step, next.score, next.best_score, &last_districts)
-                .unwrap();
-        }
-        next = recv.recv().unwrap();
-    }
-    writer.close().unwrap();
-}
-
-/// Sends the next canonical chain state to every worker.
-///
-/// # Arguments
-///
-/// * `job_sends` - Worker job channels.
-/// * `diff` - Accepted proposal to broadcast, or `None` for an all-self-loop round.
-/// * `score` - Current canonical chain score.
-fn send_tilted_jobs(
-    job_sends: &[Sender<TiltedJobPacket>],
-    diff: Option<&RecomProposal>,
-    score: f64,
-) {
-    for job in job_sends.iter() {
-        job.send(TiltedJobPacket {
-            diff: diff.cloned(),
-            current_score: score,
-            terminate: false,
-        })
-        .unwrap();
-    }
-}
-
-/// Stops all tilted-run worker threads.
-///
-/// # Arguments
-///
-/// * `job_sends` - Worker job channels.
-fn stop_tilted_workers(job_sends: &[Sender<TiltedJobPacket>]) {
-    for job in job_sends.iter() {
-        job.send(TiltedJobPacket {
-            diff: None,
-            current_score: 0.0,
-            terminate: true,
-        })
-        .unwrap();
-    }
-}
-
-/// Blocks until each worker has returned exactly one round of tilted output.
-///
-/// # Arguments
-///
-/// * `result_recv` - Shared result channel from all workers.
-/// * `n_threads` - Number of worker packets to collect.
-///
-/// # Returns
-///
-/// The total number of tilted rejections and all worker-accepted proposals for
-/// this round.
-fn collect_tilted_results(
-    result_recv: &Receiver<TiltedResultPacket>,
-    n_threads: usize,
-) -> (usize, Vec<ScoredProposal>) {
-    let mut rejections = 0;
-    let mut proposals = Vec::<ScoredProposal>::new();
-
-    for _ in 0..n_threads {
-        let packet = result_recv.recv().unwrap();
-        rejections += packet.rejections;
-        proposals.extend(packet.proposals);
-    }
-
-    (rejections, proposals)
 }
 
 /// Interleaves one batch of worker rejections/proposals into the sequential chain.
